service/src/main/java/com/shopzilla/service/shoppingcart/resource/IndexResource.java
/**
 * Copyright (C) 2004 - 2013 Shopzilla, Inc. 
 * All rights reserved. Unauthorized disclosure or distribution is prohibited.
 */
package com.shopzilla.service.shoppingcart.resource;

import com.yammer.dropwizard.views.View;

import javax.ws.rs.GET;
import javax.ws.rs.Path;
import javax.ws.rs.Produces;
import javax.ws.rs.core.MediaType;

/**
 * Index page controller.
 */
@Path("/")
@Produces(MediaType.TEXT_HTML)
public class IndexResource {

    @GET
    public IndexView handle() {
        return new IndexView("/index.mustache");
    }

    public static class IndexView extends View {

        protected IndexView(String templateName) {
            super(templateName);
        }

    }
}
